Refactor connection class to its own file.
